Move excerpt semantic token helpers into partial
